fix: SnapToValidCompressionLevel の int.MinValue オーバーフローを修正
Math.Abs(level - valid) は level=int.MinValue のとき OverflowException を
投げる。settings.json 改竄値 ZipCompressionLevel=-2147483648 が来ると
SanitizeAfterLoad の途中で吐き出して Load() の catch に拾われ、せっかくの
load 時防御がバイパスされて (CompressionLevel)int.MinValue が native 7z.dll
に到達してしまう (gemini / CodeRabbit / Codex の 3 bot 同調指摘)。

距離計算を (long) にキャストして OverflowException を回避。
int.MinValue / int.MaxValue 両境界の回帰テストを SnapTo / SanitizeAfterLoad
両 Theory に追加。

+++ src/Lhamiel/Util/Settings.cs
@@ -611,11 +611,16 @@ public class Settings
     /// </summary>
     internal static int SnapToValidCompressionLevel(int level)
     {
+        // (long) キャストして距離計算する: level が int.MinValue のとき
+        // `level - 0` の int 結果は int.MinValue で、Math.Abs(int.MinValue) は
+        // OverflowException を投げる。改竄された settings.json を Load() の catch
+        // が SanitizeAfterLoad の途中で吐き出して、せっかくの load 時防御が
+        // バイパスされて (CompressionLevel)int.MinValue が native 7z.dll に届いてしまう。
         var nearest = ValidCompressionLevels[0];
-        var bestDistance = Math.Abs(level - nearest);
+        var bestDistance = Math.Abs((long)level - nearest);
         foreach (var valid in ValidCompressionLevels)
         {
-            var distance = Math.Abs(level - valid);
+            var distance = Math.Abs((long)level - valid);
             if (distance < bestDistance)
             {
                 bestDistance = distance;
